+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.0.15.RELEASE'
 }
 
-group = 'com.example'
+group = 'com.dunji'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '11'
 
@@ -26,7 +26,7 @@ jar {
 dependencies {
     //spring boot web, JPA, Spring Security
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
+//    implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
 
@@ -37,7 +37,7 @@ dependencies {
     //h2
     runtimeOnly 'com.h2database:h2'
     //mysql
-    runtimeOnly 'com.mysql:mysql-connector-j'
+//    runtimeOnly 'com.mysql:mysql-connector-java'
 
     //gson
     implementation 'com.google.code.gson:gson:2.10'
