+++ src/main/java/com/dunji/backend/DunJiBackEndApplication.java
@@ -1,4 +1,4 @@
-package com.example.dunjibackend;
+package com.dunji.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
