+++ src/test/java/com/dunji/backend/DunJiBackEndApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.dunjibackend;
+package com.dunji.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
